Split off the sqlite statement class to its own header
